+++ apps/api/app/shared/infrastructure/db/models/portfolio_ledger.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import CHAR, ForeignKey, Numeric, Text, text
 from sqlalchemy.dialects.postgresql import TIMESTAMP, UUID
@@ -26,10 +27,10 @@ class PortfolioLedgerEntry(Base):
     kind: Mapped[str] = mapped_column(
         pg_enum(LedgerEntryKind, "ledger_entry_kind"), nullable=False
     )
-    amount_source: Mapped[float] = mapped_column(Numeric(16, 2), nullable=False)
+    amount_source: Mapped[Decimal] = mapped_column(Numeric(16, 2), nullable=False)
     currency: Mapped[str] = mapped_column(CHAR(3), nullable=False)
-    amount_eur: Mapped[float] = mapped_column(Numeric(16, 2), nullable=False)
-    rate_to_eur: Mapped[float] = mapped_column(Numeric(24, 12), nullable=False)
+    amount_eur: Mapped[Decimal] = mapped_column(Numeric(16, 2), nullable=False)
+    rate_to_eur: Mapped[Decimal] = mapped_column(Numeric(24, 12), nullable=False)
     fx_rate_at: Mapped[datetime] = mapped_column(
         TIMESTAMP(timezone=True), nullable=False
     )
